+++ CHANGELOG.md
@@ -5,3 +5,14 @@
 ## 1.1.0
 
 - Dependency update due to the zeytin_local_storage update.
+
+## 1.2.0
+
+- The ZeytinX class has been developed.
+- Database connection points have been created.
+
+You can review: src/database.dart && utils/operation.dart
+
+## 1.2.1
+
+- License Changed